pair tests to pair folder
