+++ dex/pair/Cargo.toml
@@ -39,3 +39,6 @@ features = ["esdt-token-payment-legacy-decode",]
 
 [dev-dependencies.elrond-wasm-debug]
 version = "=0.36.1"
+
+[dev-dependencies]
+num-bigint = "0.4.2"
